Extensible AppendExecutionProvider and expose OrtSessionOptions::AddConfigEntry directly (#1384)
Also simplify our AppendExecutionProvider usage. I removed the
unnecessary extra step of copying into an unordered_map then back into a
std::vector.

This also unifies all AppendExecutionProviders together and is now
extensible to future unknown providers.

Now for unknown providers, Onnxruntime will report an error with a list
of the valid providers. In testing, the list in onnxruntime was found to
be incomplete, this PR addresses this:
https://github.com/microsoft/onnxruntime/pull/24352