Changes how the device OrtAllocators work, use a global OrtSession instead (#1378)
Previously the code would store the OrtAllocator from the first
OrtSession created. This apparently didn't work as expected. Even though
the API to get an allocator from Ort is called CreateAllocator, it's
really just returning a proxy to the OrtSession's allocator. When the
OrtSession is destroyed, the OrtAllocator is invalid.

---------

Co-authored-by: Scott McKay <Scott.McKay@microsoft.com>