Fix LM head for Gemma-2 (#1420)
### Description

This PR fixes a bug with the language modeling head (LM head) for
Gemma-2.

### Motivation and Context

When Gemma-3 was added, Gemma-2 was refactored to work for both Gemma-2
and Gemma-3. During the refactoring process, however, the setting of the
`scale` attribute for the LM head of Gemma-2 was not removed. This
causes an extra Mul node to be created. It also causes two Mul nodes to
have the same name (`/lm_head/Mul`) as well.