Trim genai version for dev runs in Nuget packaging (#1538)
Trim genai version for dev runs in Nuget packaging step and fixes the
issue for Nuget packaging output for dev runs.

Currently the genai version for dev runs is something like this
"0.9.0-dev-423423-da43-d34*" and because of that it doesn't match the
file names. We just need to extract the first two parts in the genai
version and this PR does the same. Have started test runs to ensure it
works properly.