Append eos token to the end of input sequence for marian models (#1630)
SPM encoding that is used for marian model append eos token by default
after every prompt. To make that work we need to set add_special_tokens
to true but we cannot do in our case now. So we need to find a work
around to append eos.
Option 1: Append eos in the client code. But the user does not want to
do that
Option 2: Append </s> after the prompt --> This can be a back up option,
but
Option 3: Set add_special_tokens to true. --> Can discuss on this

Current change appends 0 to the end of input sequence which works as
eos.