Fixes all packaging pipelines (#1641)
Recent pull-requests broke the packaging pipeline. This pull-request
addresses the issues.

I also discovered that the binaries were no longer getting correctly
signed. The primary reason is that our `python build.py ` or `cmake
--preset ... --build` script does two things:

- Build the binaries
- Package them (into *.zip, *.tar.gz or *.whl)

As a result, there is no room to codesign the binaries (since
codesigning is an intermediate step and we cannot yield from a python
script before continuing).
In order to codesign the binaries, we need to break down the build
process in two phases.

1. Build the biaries
2. -- intermediate step to codesign --
3. Build the package

This pr introduces that change as well.