Update the benchmark scripts to account for the time spent in `generator.is_done()`.
This adjustment is crucial for the CUDA Execution Provider (EP), which relies on GPU-based sampling. In the case of CUDA EP, the generation time currently only includes the sampling kernels' launch time, not the execution time, due to synchronization occurring during the `generator.is_done()` call. For other EPs that perform sampling on the CPU, the token generation time already encompasses the sampling time.