Fix BF16 CUDA version of OpenAI's gpt-oss (#1706)
### Description

This PR fixes a bug with generating the BF16 CUDA version of OpenAI's
`gpt-oss` models.

### Motivation and Context

The shape of the MLP 1 weights in the MoE layer is different between the
OpenAI and Hugging Face implementations. OpenAI's implementation has the
following shape: `(32, 5760, 2880)`. Hugging Face's implementation has
the following shape: `(32, 2880, 5760)`. The [original PR
changes](https://github.com/microsoft/onnxruntime-genai/pull/1678) were
made with the OpenAI implementation so a transpose is inserted to fix
this gap.