fix pad vs. eos token misidentification (#1694)
Addresses this issue: #1682 regarding the misidentification of eos
tokens as padding tokens. For models which share EOS and pad token, each
EOS token present in the input would be identified as a pad token,
leading to incorrect attention_mask and later parity issues. This PR
fixes this issue by adding a special case for batch_size 1 (where input
padding is unnecessary and assumed to not be there) as well as iterating
through the input backwards in batch_size > 1 scenarios, as to not
misinterpret any EOS tokens which are not at the end of the sequence.