Use updated Decoder API with `skip_special_tokens` (#1722)
### Updates

Inside the foundry catalog for [Foundry
Local](https://github.com/[microsoft/Foundry-Local](https://github.com/microsoft/Foundry-Local)),
all tool call tokens are marked as special such that we can use them
with guidance inside
[GenAI](https://github.com/microsoft/onnxruntime-genai). Now, in order
for decoding to work with these tokens, we need to expose using special
tokens as an option in our Decoder API in Extensions. Hence, ORT
Extensions now adds `skip_special_tokens` to
`OrtxDetokenize1DWithOptions` in this
[PR](https://github.com/microsoft/onnxruntime-extensions/pull/989) which
this PR leverages in the GenAI `Decoder`.

Note that we leave `OrtxDetokenize1D` as is for backward compatibility.
In the future, we shall be adding an extensible options map API into ORT
Extensions that will reduce the need for messy parameter additions to
our Encoder/Decoder APIs, and can be easily consumed by GenAI and other
partners.

### Validation
- [x] Native C++ testing in ORT Extensions against HuggingFace
benchmark.

Co-authored-by: Sayan Shaw <sayanshaw@microsoft.com>