Update Extensions commit for Qwen 2.5 Chat Template Tools Fix (#1730)
### Updates
For Qwen, and some other models, `tools` should be a list of functions
stored in JSON format, passed in separately, as an additional input to
`apply_chat_template`. These tools were not handled correctly in the C++
backend in ApplyChatTemplate in ORT Extensions and were leading to extra
newlines. This is now fixed by this
[PR](https://github.com/microsoft/onnxruntime-extensions/pull/992).

### Validation
- [x] Python unit testing against HuggingFace.

Co-authored-by: Sayan Shaw <sayanshaw@microsoft.com>