Fix new/delete override and Enable cuda kernel test in Windows (#1772)
Previously, cuda kernel test was disabled in Windows CI since test
failed to start.
It is caused by new/delete override, which has assumption that the genai
object is created before any other object. The global initialization
order depends on compiler.

In Windows, when new is called, the genai object might not be created
yet.

Here we change the implementation of the override to fix such potential
crash in Windows.