Add trust remote code option to model builder (#1766)
### Description

This PR adds `hf_remote` to the `extra_options` to allow users to
control trusting remote code in the model builder.

### Motivation and Context

For models where the modeling code is only available in the Hugging Face
repo, this should always be true. For models where the modeling code is
in both the repo and in Hugging Face's `transformers`, this can be set
to false.