ADO Pipeline - nuget_winml_package_reference_version is configured at build time. (#1811)
`ort_nuget_winml_package_reference_version` informs the package
reference version requirements for `Microsoft.WindowsAppSDK.ML`.

Attention should be paid when providing an 'open range' for future
versions, as they could include newer ORT versions that are not
backwards compatible. It's better to be consersitive in the range at the
cost of more GenAI releases, vs. customers having runtime issues due to
a breaking ORT change that the range 'promised' was valid.

Example: `[1.8.2001,1.8.3000] ` - This range was based on the convention
that ORT would NOT change in the revisions.