Support pre-registered plug-in cuda execution provider library (#1850)
Up until now, for the cuda execution provider, onnxruntime-genai tried
using the built-in cuda ep using the legacy `OrtCUDAProviderOptionsV2`
and `AppendExecutionProvider_CUDA_V2` API. These functions are designed
to make use of the built-in cuda execution provider and are not
compatible with the world of plug-in eps.

The changes in this pull-request now extend support to executing the
model with a pre-registered (plugged-in) CUDAExecutionProvider.

In order to use the plug-in capabilities, application layer needs to do
the following:

- C++

  ```cpp
  #include "onnxruntime_cxx_api.h"
  #include "ort_genai.h"
  
  auto env = Ort::Env();
env.RegisterExecutionProviderLibrary("CUDAExecutionProvider",
"path\to\onnxruntime_providers_cuda.dll");
  
  auto model = OgaModel("path\to\model\directory");
  ...
  ```

- Python

  ```python
  import onnxruntime_genai as og
  
og.register_execution_provider_library("CUDAExecutionProvider",
"path\to\onnxruntime_providers_cuda.dll")
  
  model = og.Model("path\to\model\directory")
  ...
  ```

- C#

  ```csharp
  using Microsoft.ML.OnnxRuntime;
  using Microsoft.ML.OnnxRuntimeGenAI;
  
  var ortEnv = OrtEnv.Instance();
ortEnv.RegisterExecutionProviderLibrary("CUDAExecutionProvider",
"path\to\onnxruntime_providers_cuda.dll");
  
  using Model model = new Model("path\to\model\directory");
  ...
  ```