Fix handling EOS token id detection (#1849)
### Description

This PR fixes how EOS token id detection is handled with ONNX Runtime
GenAI when generating tokens. An extra check after generating the next
token resolves the issue.

### Motivation and Context

A PR was [recently
made](https://github.com/microsoft/onnxruntime-genai/pull/1814/) in ONNX
Runtime GenAI to omit inserting the EOS token id in the ongoing sequence
of token ids during continuous decoding. Previously, the EOS token id
would be appended to the sequence of token ids that are generated. The
user could see this when calling the `GetSequence` API.

That behavior is not good for continuous decoding because an EOS token
id indicates the conversation has ended. Appending subsequent token ids
after an EOS token id can result in undefined behavior.

Although the PR fixes this, the PR changes how the default behavior of
the generation loop works in ORT GenAI. The current behavior is as
follows.

```
while not IsDone():
    GenerateToken()
    GetLastToken()
    PrintLastToken()
```

When `GenerateToken` produces the token id before EOS, `GetLastToken`
will return that token id. In the next iteration of the `while` loop,
`GenerateToken` produces the EOS token id. But since the EOS token id is
not added to the sequence of generated token ids, `GetLastToken` still
returns the token id before EOS. Thus, the token id before EOS is
printed twice.

This can be resolved by checking `IsDone` after `GenerateToken` and exit
the `while` loop early.

```
while not IsDone():
    GenerateToken()
    if IsDone():
        break
    GetLastToken()
    PrintLastToken()
```

This can be simplified further by changing the `while` loop to an
infinite loop.

```
while True:
    GenerateToken()
    if IsDone():
        break
    GetLastToken()
    PrintLastToken()
```