Ensure Consistent Tool Calling JSON Serialization and Deserialization (#1863)
### Description

This PR updates the tool normalization changes by addressing the issue
of inconsistent handling of tool quotes in Phi-4 function calling and
maintains consistency in JSON serialization and deserialization,
particularly when tools are embedded within the message body as it is in
the case of Phi-4, rather than as an additional input like with Qwen 2.5
and most models.

### Motivation and Context

These changes were added to ORT extensions in [this
PR](https://github.com/microsoft/onnxruntime-extensions/pull/1006).

Co-authored-by: Sayan Shaw <sayanshaw@microsoft.com>