Add C# binding for GetNextTokens (#1865)
### Description

This PR adds `GetNextTokens` as an API to the C# bindings.

### Motivation and Context

Currently in C#, the user has to get the entire sequence first and then
access the last element of the sequence to extract the next token (i.e.
`generator.GetSequence(0)[^1]`). In Python, the user just has to call
`generator.get_next_tokens()[0]` to get a list of the next tokens and
only access the first token in that list.

With this PR, the generation loop in the C# and Python examples now
looks and runs similarly.