Fix bug in Squeeze for getting the value of total_seq_len (#1886)
The `Squeeze` op is used for removing single-dimensional entries from
the shape of a tensor. In this node the `axes` input is set to `[0]`
which would only eliminate the first axis and lead to the output shape
to be `[1]` if the `batch_size` is 1.This would cause ShapeInference
error at
https://github.com/microsoft/onnxruntime/blob/main/onnxruntime/core/graph/graph.cc#L111-L132
if it is not a strict mode.

This PR fixes the issue by:
- Changing the ReduceSum output shape to [batch_size] by adding
keepdims=0 attribute
- Using ReduceMax instead of Squeeze to get the value of total_seq_len
and make it as a scalar, this would cover scenarios when batch_size > 1