whisper: Support a variant of the whisper pipeline where encoder / decoder are stateful. (#1857)
This will mainly be used to support EPCtx-wrapped OpenVINO IR's -- the
same set of models that are compatible with OpenVINO GenAI.

In this case, unlike the default pipeline, the encoder doesn't pull in
KV cross projection -- it only outputs encoder_hidden_states.

For the decoder portion of the pipeline, the cross & self KV cache
tensors are managed internal to the model (attached to the session).